Rename artifacts
After revisiting this project after some time, I find the names of
the artifacts confusing and challenging to remember.

The new nomenclature is in the form <primary dep>-<secondary dep>, so
`tomcat-log4j` are Tomcat components that help integrating Log4j
and `log4j-tomcat` are Log4j components/plugins that help integrating it
into Tomcat.